Move post onboarding types to LLC
